+++ vs/BulletStream.h
@@ -8,8 +8,8 @@
 
 class BulletStream {
 public:
-	void addBullet(Vector ship_position, float angle, float ship_width);
-	void updateBullets(float dt);
+	void addBullet(Vector ship_position, double angle, double ship_width);
+	void updateBullets(double dt);
 	void drawAll();
 	std::vector<Bullet>& getBullets();
 	void clear();
